% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Health and Demographics Dataset}
\format{
A data frame with 10,000 rows and 15 variables:
\describe{
\item{age}{Patient age in years}
\item{height}{Patient height}
\item{weight}{Patient weight}
\item{SBP}{Systolic blood pressure}
\item{cholesterol}{Cholesterol level}
\item{female}{Binary indicator for sex (1 = female, 0 = male)}
\item{history}{Family history of heart attacks indicator}
\item{smoke}{Smoking status (1 = smoker, 0 = non-smoker)}
\item{hyperten}{Hypertension status (1 = yes, 0 = no)}
\item{heart}{Heart attack indicator (1 = yes, 0 = no)}
\item{socio}{Socioeconomic status}
\item{marital}{Marital status (e.g. Married)}
\item{employ}{Employment status (e.g. Employed)}
\item{race}{Race/ethnicity (e.g. White)}
\item{edu}{Education level (e.g. BS)}
}
}
\source{
This is simulated data created by the package maintainer.
}
\usage{
data_example
}
\description{
A dataset containing 10,000 rows and 15 variables. This can
be used to practice implementing AutoTab.
}
\keyword{datasets}
