% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_distribution.R
\name{extracting_distribution}
\alias{extracting_distribution}
\title{Extracting_Distribution will create feat_dist which will be used throughout the package.
There is one row per variable. Column one, column_name, indicates the names of the variables in the dataset.
Column two, distribution, indicates the type of distribution the variable falls under.
There are three options, Gaussian, Bernoulli, and categorical.
Column three, num_params, indicates how many parameters the VAE should produce for the given variable.}
\usage{
extracting_distribution(data)
}
\arguments{
\item{data}{Your original input data, prior to pre-processing}
}
\value{
feat_dist which is used in AutoTab to define variable distribution types
}
\description{
Extracting_Distribution will create feat_dist which will be used throughout the package.
There is one row per variable. Column one, column_name, indicates the names of the variables in the dataset.
Column two, distribution, indicates the type of distribution the variable falls under.
There are three options, Gaussian, Bernoulli, and categorical.
Column three, num_params, indicates how many parameters the VAE should produce for the given variable.
}
