% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_distribution.R
\name{feat_reorder}
\alias{feat_reorder}
\title{Reorder \code{feat_dist} rows to match preprocessed data}
\usage{
feat_reorder(feat_dist, data)
}
\arguments{
\item{feat_dist}{Data frame from \code{\link[=extracting_distribution]{extracting_distribution()}}.}

\item{data}{Data frame of the \strong{original (preprocessed)} variables.}
}
\value{
The input \code{feat_dist}, reordered to align with \code{data}.
}
\description{
Ensures row order in \code{feat_dist} matches the \strong{column prefix order} in the
preprocessed (dummy-coded) training data. This assumes dummy columns are
named as \verb{<original_name>_<level>} and therefore start with the original
variable name.
}
\examples{
\dontrun{
feat_dist = feat_reorder(extracting_distribution(data_final),data)
rownames(feat_dist) = NULL
set_feat_dist(feat_dist)
}
}
\seealso{
\code{\link[=extracting_distribution]{extracting_distribution()}}, \code{\link[=set_feat_dist]{set_feat_dist()}}
}
