% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_dist_storinginternally.R
\name{get_feat_dist}
\alias{get_feat_dist}
\title{Get the stored feature distribution}
\usage{
get_feat_dist()
}
\value{
A data.frame containing feature distribution metadata.
}
\description{
Retrieves the \code{feat_dist} object previously stored by
\code{set_feat_dist()}.  Throws an error if it has not been set.
}
