% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_distribution.R
\name{feat_reorder}
\alias{feat_reorder}
\title{If feat_dist is misordered, the function feat_reorder can be applied.
The user wants the rows of feat_order to match the order of the pre-processed dataset that will be used for training the VAE.
Feat_order ensures the row order in feat_dist matches the order in the input dataset.}
\usage{
feat_reorder(feat_dist, data)
}
\arguments{
\item{feat_dist, data}{feat_dist is the output of extracting_distribution. Data is your pre-processed data.}
}
\value{
feat_dist that is in the same order as the pre-processed data going into the VAE
}
\description{
If feat_dist is misordered, the function feat_reorder can be applied.
The user wants the rows of feat_order to match the order of the pre-processed dataset that will be used for training the VAE.
Feat_order ensures the row order in feat_dist matches the order in the input dataset.
}
