% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_sampling.R
\name{Latent_sample}
\alias{Latent_sample}
\title{Sample from the latent space using the reparameterization trick}
\usage{
Latent_sample(z_mean, z_log_var)
}
\arguments{
\item{z_mean, z_log_var}{The params of the latent space.}
}
\value{
A sample from the latent space.
}
\description{
Sample from the latent space using the reparameterization trick
}
