% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{VAE_train}
\alias{VAE_train}
\title{This will call your VAE training. Ensure you have your encoder_info and decoder_info.}
\usage{
VAE_train(
  data,
  encoder_info,
  decoder_info,
  Lip_en,
  pi_enc = 1,
  lip_dec,
  pi_dec = 1,
  latent_dim,
  epoch,
  beta,
  kl_warm = FALSE,
  kl_cyclical = FALSE,
  n_cycles,
  ratio,
  beta_epoch = 15,
  temperature,
  temp_warm = FALSE,
  temp_epoch,
  batchsize,
  wait,
  min_delta,
  lr,
  max_std = 10,
  min_val,
  weighted = 0,
  recon_weights,
  seperate = 0,
  prior = "single_gaussian",
  K = 3,
  learnable_mog = FALSE,
  mog_means = NULL,
  mog_log_vars = NULL,
  mog_weights = NULL
)
}
\value{
your training step
}
\description{
This will call your VAE training. Ensure you have your encoder_info and decoder_info.
}
